Eclipse source clean up

+++ src/main/java/com/firemerald/additionalplacements/block/interfaces/IStairBlock.java
@@ -6,15 +6,15 @@ import java.util.function.Function;
 
 import javax.annotation.Nullable;
 
-import com.firemerald.additionalplacements.generation.APGenerationTypes;
-import com.firemerald.additionalplacements.generation.GenerationType;
-import com.firemerald.additionalplacements.util.*;
-import com.firemerald.additionalplacements.block.stairs.StairConnectionsType;
 import com.firemerald.additionalplacements.block.stairs.AdditionalStairBlock;
+import com.firemerald.additionalplacements.block.stairs.StairConnectionsType;
 import com.firemerald.additionalplacements.block.stairs.common.CommonStairShape;
 import com.firemerald.additionalplacements.block.stairs.common.CommonStairShapeState;
 import com.firemerald.additionalplacements.block.stairs.vanilla.VanillaStairShapeState;
 import com.firemerald.additionalplacements.client.BlockHighlightHelper;
+import com.firemerald.additionalplacements.generation.APGenerationTypes;
+import com.firemerald.additionalplacements.generation.GenerationType;
+import com.firemerald.additionalplacements.util.ComplexFacing;
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.blaze3d.vertex.VertexConsumer;
 import com.mojang.math.Matrix3f;
@@ -60,7 +60,7 @@ public interface IStairBlock<T extends Block> extends IPlacementBlock<T>
 		else
 			return getBlockStateInternal(shapeState, currentState);
 	}
-	
+
 	public BlockState getBlockStateInternal(CommonStairShapeState shapeState, BlockState currentState);
 
 	@Override
@@ -71,13 +71,13 @@ public interface IStairBlock<T extends Block> extends IPlacementBlock<T>
 		ComplexFacing newFacing = ComplexFacing.forFacing(transform.apply(oldFacing.forward), transform.apply(oldFacing.up));
 		return getBlockState(newFacing, state.shape, blockState);
 	}
-	
+
 	public static ComplexFacing getFacingOrNull(BlockState blockState) {
 		return blockState.getBlock() instanceof IStairBlock ?  ((IStairBlock<?>) blockState.getBlock()).getShapeState(blockState).facing : null;
 	}
 
 	public abstract CommonStairShapeState getShapeState(BlockState blockState);
-	
+
 	public StairConnectionsType connectionsType();
 
 	@Override
@@ -204,14 +204,14 @@ public interface IStairBlock<T extends Block> extends IPlacementBlock<T>
 	public static final float ARROW_OFFSET = -0.4375f;
 	public static final float ARROW_OUTER = 0.375f;
 	public static final float ARROW_INNER = 0.125f;
-	
+
 	@Override
 	@Environment(EnvType.CLIENT)
 	public default void renderPlacementPreview(PoseStack pose, VertexConsumer vertexConsumer, Player player, BlockHitResult result, float partial, float r, float g, float b, float a) {
 		if (!this.connectionsType().allowFlipped) return;
-		ComplexFacing facing = getFacing(result.getDirection(), 
-				(float) (result.getLocation().x - result.getBlockPos().getX() - .5), 
-				(float) (result.getLocation().y - result.getBlockPos().getY() - .5), 
+		ComplexFacing facing = getFacing(result.getDirection(),
+				(float) (result.getLocation().x - result.getBlockPos().getX() - .5),
+				(float) (result.getLocation().y - result.getBlockPos().getY() - .5),
 				(float) (result.getLocation().z - result.getBlockPos().getZ() - .5));
 		//z is up
 		//y is forward
@@ -222,12 +222,12 @@ public interface IStairBlock<T extends Block> extends IPlacementBlock<T>
 				facing.right  .getStepX(), facing.right  .getStepY(), facing.right  .getStepZ(), 0,
 				facing.forward.getStepX(), facing.forward.getStepY(), facing.forward.getStepZ(), 0,
 				facing.up     .getStepX(), facing.up     .getStepY(), facing.up     .getStepZ(), 0,
-				0, 0, 0, 1 
+				0, 0, 0, 1
 				}));
 		pose.mulPoseMatrix(mat);
 		Matrix4f poseMat = pose.last().pose();
 		Matrix3f normMat = pose.last().normal();
-		BlockHighlightHelper.lineLoop(vertexConsumer, poseMat, normMat, ARROW_OFFSET, r, g, b, a, 
+		BlockHighlightHelper.lineLoop(vertexConsumer, poseMat, normMat, ARROW_OFFSET, r, g, b, a,
 				 0          ,  ARROW_OUTER,
 				 ARROW_OUTER,  0          ,
 				 ARROW_INNER,  0          ,
@@ -247,31 +247,31 @@ public interface IStairBlock<T extends Block> extends IPlacementBlock<T>
 	{
 		Matrix4f poseMat = pose.last().pose();
 		Matrix3f normMat = pose.last().normal();
-		
+
 		//outer box
-		BlockHighlightHelper.lineCenteredSquare(vertexConsumer, poseMat, normMat, -OUTER_EDGE, r, g, b, a, 
+		BlockHighlightHelper.lineCenteredSquare(vertexConsumer, poseMat, normMat, -OUTER_EDGE, r, g, b, a,
 				OUTER_EDGE);
-		
+
 		if (this.connectionsType().allowFlipped) {
 			//inner edges
-			BlockHighlightHelper.lineCenteredSquare(vertexConsumer, poseMat, normMat, -OUTER_EDGE, r, g, b, a, 
+			BlockHighlightHelper.lineCenteredSquare(vertexConsumer, poseMat, normMat, -OUTER_EDGE, r, g, b, a,
 					OUTER_EDGE);
-			
+
 			//middle cross
-			BlockHighlightHelper.lineCenteredCross(vertexConsumer, poseMat, normMat, -OUTER_EDGE, r, g, b, a, 
+			BlockHighlightHelper.lineCenteredCross(vertexConsumer, poseMat, normMat, -OUTER_EDGE, r, g, b, a,
 					OUTER_EDGE);
 		} else {
 			//corners
-			BlockHighlightHelper.lineOctal(vertexConsumer, poseMat, normMat, -OUTER_EDGE, r, g, b, a, 
+			BlockHighlightHelper.lineOctal(vertexConsumer, poseMat, normMat, -OUTER_EDGE, r, g, b, a,
 					INNER_EDGE, INNER_EDGE,
 					OUTER_EDGE, INNER_EDGE);
-			
+
 			//middle cross
-			BlockHighlightHelper.lineCenteredCross(vertexConsumer, poseMat, normMat, -OUTER_EDGE, r, g, b, a, 
+			BlockHighlightHelper.lineCenteredCross(vertexConsumer, poseMat, normMat, -OUTER_EDGE, r, g, b, a,
 					INNER_EDGE);
 		}
 	}
-    
+
 	@Override
 	public default GenerationType<?, ?> getGenerationType() {
 		return APGenerationTypes.stairs();
@@ -289,9 +289,9 @@ public interface IStairBlock<T extends Block> extends IPlacementBlock<T>
 
 	public default ComplexFacing getFacing(Direction out, Vec3 hitPos, Vec3i blockPos)
 	{
-		return getFacing(out, 
-				(float) (hitPos.x - blockPos.getX() - .5), 
-				(float) (hitPos.y - blockPos.getY() - .5), 
+		return getFacing(out,
+				(float) (hitPos.x - blockPos.getX() - .5),
+				(float) (hitPos.y - blockPos.getY() - .5),
 				(float) (hitPos.z - blockPos.getZ() - .5));
 	}
 
@@ -299,16 +299,16 @@ public interface IStairBlock<T extends Block> extends IPlacementBlock<T>
 	{
 		switch (out.getAxis()) {
 		case X:
-			return getFacingFromSide((float) hitZ, (float) hitY, Direction.SOUTH, Direction.UP, out);
+			return getFacingFromSide(hitZ, hitY, Direction.SOUTH, Direction.UP, out);
 		case Y:
-			return getFacingFromSide((float) hitX, (float) hitZ, Direction.EAST, Direction.SOUTH, out);
+			return getFacingFromSide(hitX, hitZ, Direction.EAST, Direction.SOUTH, out);
 		case Z:
-			return getFacingFromSide((float) hitX, (float) hitY, Direction.EAST, Direction.UP, out);
-		default: 
+			return getFacingFromSide(hitX, hitY, Direction.EAST, Direction.UP, out);
+		default:
 			return ComplexFacing.SOUTH_UP;
 		}
 	}
-	
+
 	public default ComplexFacing getFacingFromSide(float localX, float localY, Direction localRight, Direction localUp, Direction localOut) {
 		if (localY > localX) { //top-left half
 			if (localY > -localX) { //top quarter
@@ -324,7 +324,7 @@ public interface IStairBlock<T extends Block> extends IPlacementBlock<T>
 			}
 		}
 	}
-	
+
 	public default ComplexFacing getFacingFromQuarter(float localX, float localY, Direction localRight, Direction localUp, Direction localOut) {
 		Direction forward, up;
 		if (localY > INNER_EDGE) { //top half
